Primary keys Long -> Int Not very good idea use long

+++ src/main/java/com/megabank/backend/service/dao/AccountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
-	List<Account> findAllByUserId(long userId);
-	Account findByUserIdAndId(long userId, long accountId);
+public interface AccountRepository extends JpaRepository<Account, Integer> {
+	List<Account> findAllByUserId(int userId);
+	Account findByUserIdAndId(int userId, int accountId);
 }
